# Conflicts:
#	.idea/.idea.backend/.idea/workspace.xml
#	appsettings.Development.json
#	bin/Debug/net9.0/appsettings.Development.json
#	bin/Debug/net9.0/backend.dll
#	bin/Debug/net9.0/backend.exe
#	bin/Debug/net9.0/backend.pdb
#	obj/Debug/net9.0/apphost.exe
#	obj/Debug/net9.0/backend.AssemblyInfo.cs
#	obj/Debug/net9.0/backend.AssemblyInfoInputs.cache
#	obj/Debug/net9.0/backend.csproj.CoreCompileInputs.cache
#	obj/Debug/net9.0/backend.dll
#	obj/Debug/net9.0/backend.pdb
#	obj/Debug/net9.0/backend.sourcelink.json
#	obj/Debug/net9.0/ref/backend.dll
#	obj/Debug/net9.0/refint/backend.dll
#	obj/rider.project.model.nuget.info
#	obj/rider.project.restore.info

+++ Renting/Domain/Repositories/IRentRepository.cs
@@ -7,6 +7,6 @@ public interface IRentRepository : IBaseRepository<Rent>
 {
     Task<Rent?> GetByIdAsync(int id);
     
-    Task<Rent?> GetByUserId(long userId);
+    Task<Rent?> GetByUserId(int userId);
     
 }
